feat(sdk-v4): clean up dead code and improve readme
