+++ packages/sdk/test/bundle.test.ts
@@ -2,10 +2,7 @@
  * @jest-environment node
  */
 
-const {
-  createMoneriumApiClient,
-  clientCredentialsGrant,
-} = require('../dist/index.js');
+const { MoneriumPrivateClient } = require('../dist/index.js');
 
 import * as constants from './constants';
 
@@ -16,21 +13,24 @@ const {
 } = constants;
 
 test('should import without throwing', () => {
-  expect(createMoneriumApiClient).toBeDefined();
-  expect(clientCredentialsGrant).toBeDefined();
+  expect(MoneriumPrivateClient).toBeDefined();
 });
 
 process.env.CI !== 'true' &&
   test('CommonJs bundle smoke test', async () => {
-    const { access_token } = await clientCredentialsGrant({
+    let client = new MoneriumPrivateClient({
       environment: 'sandbox',
-      clientId: APP_ONE_CREDENTIALS_CLIENT_ID,
-      clientSecret: APP_ONE_CREDENTIALS_SECRET,
+      getAccessToken: async () => undefined,
     });
 
-    const client = createMoneriumApiClient({
+    const { access_token } = await client.clientCredentialsGrant(
+      APP_ONE_CREDENTIALS_CLIENT_ID,
+      APP_ONE_CREDENTIALS_SECRET
+    );
+
+    client = new MoneriumPrivateClient({
       environment: 'sandbox',
-      accessToken: access_token,
+      getAccessToken: async () => access_token,
     });
 
     const { profiles } = await client.getProfiles();
